Init feature sliced
